refactor(route): Move info route handler to controller file
close #329
